+++ frontend/lost/src/containers/Users/GroupsTable.jsx
@@ -43,6 +43,10 @@ export const Groups = () => {
                             Header: 'Group',
                             accessor: 'name',
                         },
+                        {
+                            Header: 'ID',
+                            accessor: 'idx',
+                        },
                         // Note: commented this out temporarily since it allows you to delete groups assigned to users
                         // The backend api needs to be fixed before re-enabling this feature
                         // {